requirements.txt
pymongo==4.12.0
fastapi==0.115.12
motor==3.7.0
psutil==7.0.0
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.1.0
uvicorn==0.34.1
docling==2.30.0
accelerate==1.6.0